Move constraint generation into its own module
_qspec_reconcile.py now contains only annotate_via_reconciliation
(the drain-loop driver). Everything else that used to live there --
the _AnnotationContext bundle, _generate_constraints_for_node,
_adjacent_edge_constraints, _shared_state_constraints,
_find_input_slot_for_producer, _nodes_covered_by -- moves to a new
_qspec_constraint_generation module.

The two context/coverage symbols quantizer.py imports (_AnnotationContext,
_nodes_covered_by) are re-exported from _qspec_reconcile so no call
sites move.
